proyecto6/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
flask==3.0.0
requests==2.31.0
sendgrid==6.10.0
python-dotenv==1.0.0
jinja2==3.1.2
email-validator==2.1.0.post1
